+++ src/main/java/com/codetaylor/mc/pyrotech/modules/storage/tile/spi/TileBagBase.java
@@ -99,6 +99,8 @@ public abstract class TileBagBase
 
   public abstract int getItemCapacity();
 
+  public abstract String getRockFillTextureLocation();
+
   public void setOpen(boolean open) {
 
     if (open && !this.isOpen()) {